+++ src/AppRoute.js
@@ -2,9 +2,9 @@ import React from 'react';
 import { BrowserRouter, Route,Router, Routes} from 'react-router-dom';
 
 
-import Splash from './Splash.js'
-import Login from './Login.js'
-import ProfilePage from './Profile.js'
+import Splash from './pages/Splash.js'
+import Login from './pages/Login.js'
+import ProfilePage from './pages/Profile.js'
 import App from './App.js'
 import Sidebar from './components/SideBar.js'
 function AppRouter() {
